refactor: 리포지토리 패키지 이름 축약(이름 뒤의 repository 제거)
