+++ src/main/java/com/codingchosun/backend/controller/ImageController.java
@@ -5,13 +5,11 @@ import com.codingchosun.backend.domain.User;
 import com.codingchosun.backend.exception.LoggedInUserNotFound;
 import com.codingchosun.backend.exception.invalidrequest.IsNotPostAuthor;
 import com.codingchosun.backend.exception.notfoundfromdb.PostNotFoundFromDB;
-import com.codingchosun.backend.repository.postrepository.DataJpaPostRepository;
-import com.codingchosun.backend.repository.userrepository.DataJpaUserRepository;
+import com.codingchosun.backend.repository.post.DataJpaPostRepository;
+import com.codingchosun.backend.repository.user.DataJpaUserRepository;
 import com.codingchosun.backend.request.ImageDeleteRequest;
 import com.codingchosun.backend.response.ApiResponse;
 import com.codingchosun.backend.service.ImageService;
-import com.codingchosun.backend.service.PostService;
-import com.codingchosun.backend.web.argumentresolver.Login;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
