+++ src/main/java/com/codingchosun/backend/repository/hashtag/DataJpaPostHashRepository.java
@@ -1,4 +1,4 @@
-package com.codingchosun.backend.repository.hashtagrepository;
+package com.codingchosun.backend.repository.hashtag;
 
 import com.codingchosun.backend.domain.Hashtag;
 import com.codingchosun.backend.domain.Post;
@@ -9,10 +9,15 @@ import java.util.List;
 import java.util.Optional;
 
 public interface DataJpaPostHashRepository extends JpaRepository<PostHash, Long> {
+
     List<PostHash> findAllByPost_PostId(Long postId);
+
     List<PostHash> findAllByHashtag_HashtagId(Long hashtagId);
+
     List<PostHash> findAllByHashtag_HashtagName(String hashtagName);
+
     Optional<PostHash> findByPostAndHashtag(Post post, Hashtag hashtag);
 
     List<PostHash> findAllByHashtagIn(List<Hashtag> hashtags);
+
 }
