+++ src/main/java/com/codingchosun/backend/repository/post/DataJpaPostRepository.java
@@ -1,5 +1,4 @@
-package com.codingchosun.backend.repository.postrepository;
-
+package com.codingchosun.backend.repository.post;
 
 import com.codingchosun.backend.constants.StateCode;
 import com.codingchosun.backend.domain.Post;
@@ -7,26 +6,21 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 public interface DataJpaPostRepository extends JpaRepository<Post, Long>, DataJpaPostRepositoryCustom {
-    List<Post> findByTitle(String title);
-
-    Page<Post> findAllByOrderByCreatedAtDesc(Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.stateCode = 'ACTIVE' ORDER BY p.createdAt DESC")
     Page<Post> findAllActiveByOrderByCreatedAtDesc(Pageable pageable);
 
     List<Post> findAllByStartTimeBeforeAndStateCode(LocalDateTime startTime, StateCode stateCode);
 
-    List<Post> findAllByEndTimeBeforeAndStateCode(LocalDateTime EndTime, StateCode stateCode);
+    @Query(value = "SELECT p FROM Post p JOIN p.postUsers pu WHERE pu.user.loginId = :loginId ORDER BY p.postId DESC")
+    List<Post> findParticipatedPosts(@Param("loginId") String loginId);
 
-    Page<Post> findAllByTitleContainingAndStateCode(String title, Pageable pageable, StateCode stateCode);
-
-    List<Post> findAllByStateCodeAndTitleContaining(StateCode stateCode,String title);
-
-    Page<Post> findAllByStateCode(Pageable pageable, StateCode stateCode);
-
+    @Query("SELECT DISTINCT p FROM Post p JOIN p.postHashes ph WHERE ph.hashtag.hashtagId IN :hashtagIds ORDER BY p.createdAt DESC")
+    Page<Post> findPostsByHashTagIdIn(@Param("hashtagIds") List<Long> hashtagIds, Pageable pageable);
 }
