+++ src/main/java/com/codingchosun/backend/repository/post/DataJpaPostRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.codingchosun.backend.repository.postrepository;
+package com.codingchosun.backend.repository.post;
 
 import com.codingchosun.backend.domain.Post;
 import org.springframework.data.domain.Page;
@@ -7,7 +7,9 @@ import org.springframework.data.domain.Pageable;
 import java.util.List;
 
 public interface DataJpaPostRepositoryCustom {
+
     Page<Post> findPostsByHashTagId(List<Long> hashTagId, Pageable pageable);
 
-    Page<Post> findPostsByResearchQuery(List<String> titleQuery, List<String> hashQuery, Pageable pageable);
+    Page<Post> findPostsBySearchQuery(List<String> titleQuery, List<String> hashQuery, Pageable pageable);
+
 }
