+++ src/main/java/com/codingchosun/backend/repository/template/TemplateRepository.java
@@ -1,5 +1,4 @@
-package com.codingchosun.backend.repository.templaterepository;
-
+package com.codingchosun.backend.repository.template;
 
 import com.codingchosun.backend.domain.Template;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +7,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface TemplateRepository extends JpaRepository<Template, Long> {
+
     Optional<Template> findTemplateByContent(String content);
 
     List<Template> findByValidates_ToUser_UserId(Long id);
+
 }
