+++ src/main/java/com/codingchosun/backend/repository/user/DataJpaUserRepository.java
@@ -3,6 +3,8 @@ package com.codingchosun.backend.repository.user;
 import com.codingchosun.backend.constants.StateCode;
 import com.codingchosun.backend.domain.User;
 import jakarta.transaction.Transactional;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -18,11 +20,13 @@ public interface DataJpaUserRepository extends JpaRepository<User, Long> {
 
     User findByNickname(String nickName);
 
-    User findByUserId(long id);
+    Optional<User> findByUserId(long id);
 
     Optional<User> findByNameAndEmail(String name, String email);
 
     Optional<User> findByNameAndEmailAndLoginId(String name, String email, String loginId);
 
-    User findByStateAndLoginId(StateCode state, String loginId);
+    Optional<User> findByStateCodeAndLoginId(StateCode state, String loginId);
+
+    User getUserByLoginId(@NotNull @Size(min = 6, max = 16) String loginId);
 }
