+++ src/main/java/com/codingchosun/backend/service/ValidateService.java
@@ -5,11 +5,11 @@ import com.codingchosun.backend.domain.*;
 import com.codingchosun.backend.exception.NotEqualsUserSize;
 import com.codingchosun.backend.exception.ObjectNotFound;
 import com.codingchosun.backend.exception.invalidrequest.AlreadyValidated;
-import com.codingchosun.backend.repository.postrepository.DataJpaPostRepository;
-import com.codingchosun.backend.repository.postuserrepository.DataJpaPostUserRepository;
-import com.codingchosun.backend.repository.templaterepository.TemplateRepository;
-import com.codingchosun.backend.repository.userrepository.DataJpaUserRepository;
-import com.codingchosun.backend.repository.validaterepository.ValidateRepository;
+import com.codingchosun.backend.repository.post.DataJpaPostRepository;
+import com.codingchosun.backend.repository.postuser.DataJpaPostUserRepository;
+import com.codingchosun.backend.repository.template.TemplateRepository;
+import com.codingchosun.backend.repository.user.DataJpaUserRepository;
+import com.codingchosun.backend.repository.validate.ValidateRepository;
 import com.codingchosun.backend.request.UserValidate;
 import com.codingchosun.backend.request.UserValidationRequest;
 import com.codingchosun.backend.request.ValidateRequest;
